Centralise tracking context for both apps and web
