+++ packages/article/styles/header/shared.js
@@ -4,8 +4,7 @@ const sharedStyles = {
     lineHeight: 32,
     color: "#1d1d1b",
     marginBottom: 7,
-    fontFamily: "TimesModern-Bold",
-    fontWeight: 400
+    fontFamily: "TimesModern-Bold"
   },
   standFirst: {
     fontSize: 20,
